Moved utility functions to utility package
